rk35xx-panthor-1GHz(:10)
Original-Subject: RFC: arm64: dts: rockchip: rk3588: Use scmi gpu clk as main GPU clock
CLK_GPU is the main clock for the GPU on RK3588, it's typical source
pll can be one of gpll, cpll, aupll, npll or spll. For higher clock
rates it is also possible to use the gpu pvtpll as pll source.

The logic to switch between a normal pll and the pvtpll depending on
rate is handled in TF-A firmware, and exposed to Linux as a scmi clock.
TF-A will typically change to use normal pll for rates up to 200 MHz and
use pvtpll for 300 MHz or more.

Change to use the SCMI_CLK_GPU as the main GPU clock and add the normal
CLK_GPU as a bus clk to model this in a similar way as on RK356x.

Prior to this change the GPU clk rate was max 850 MHz:

  $ glmark2-es2-gbm -b terrain
  [...]
    GL_VENDOR:      Mesa
    GL_RENDERER:    Mali-G610 (Panfrost)
    GL_VERSION:     OpenGL ES 3.1 Mesa 25.0.4
    Surface Config: buf=32 r=8 g=8 b=8 a=8 depth=24 stencil=0 samples=0
    Surface Size:   800x600 fullscreen
  [...]
  [terrain] <default>: FPS: 139 FrameTime: 7.231 ms

After this the GPU clk rate can use the 1 GHz rate with PVTPLL:

  [terrain] <default>: FPS: 152 FrameTime: 6.579 ms

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>
X-Armbian: Patch-File: rk35xx-panthor-1GHz
X-Armbian: Patch-File-Counter: 10
X-Armbian: Patch-Rel-Directory: patch/kernel/archive/rockchip64-7.3
X-Armbian: Patch-Type: kernel
X-Armbian: Patch-Root-Type: core
X-Armbian: Patch-Sub-Type: common
X-Armbian: Original-Subject: RFC: arm64: dts: rockchip: rk3588: Use scmi gpu clk as main GPU clock